requirements.txt
appdirs==1.4.4
attrs==21.2.0
black==21.6b0
click==8.0.1
flake8==3.9.2
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pycairo==1.20.1
pycodestyle==2.7.0
pyflakes==2.3.1
PyGObject==3.40.1
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.18.0
regex==2021.4.4
toml==0.10.2
